// src/types/anilist.ts
export interface PageInfo {
  /** The total number of items. Note: This value is not guaranteed to be accurate, do not rely on this for logic*/
  total: Maybe<number>;
  /** The count on a page*/
  perPage: Maybe<number>;
  /** The current page*/
  currentPage: Maybe<number>;
  /** The last page*/
  lastPage: Maybe<number>;
  /** If there is another page*/
  hasNextPage: Maybe<boolean>;
}

export interface PageArgs {
  /** The page number*/
  page?: number;
  /** The amount of entries per page, max 50*/
  perPage?: number;
}

/** Media list watching/reading status enum. */
export enum MediaListStatus {
  Current = "CURRENT",
  Planning = "PLANNING",
  Completed = "COMPLETED",
  Dropped = "DROPPED",
  Paused = "PAUSED",
  Repeating = "REPEATING",
}

export interface MediaConnection {
  edges: Maybe<MediaEdge[]>;
  nodes: Maybe<Media[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}

/** Media connection edge */
export interface MediaEdge {
  node: Maybe<Media>;
  /** The id of the connection*/
  id: Maybe<number>;
  /** The type of relation to the parent model*/
  relationType: Maybe<MediaRelation>;
  /** If the studio is the main animation studio of the media (For Studio->MediaConnection field only)*/
  isMainStudio: boolean;
  /** The characters in the media voiced by the parent actor*/
  characters: Maybe<Character[]>;
  /** The characters role in the media*/
  characterRole: Maybe<CharacterRole>;
  /** Media specific character name*/
  characterName: Maybe<string>;
  /** Notes regarding the VA's role for the character*/
  roleNotes: Maybe<string>;
  /** Used for grouping roles where multiple dubs exist for the same language. Either dubbing company name or language variant.*/
  dubGroup: Maybe<string>;
  /** The role of the staff member in the production of the media*/
  staffRole: Maybe<string>;
  /** The voice actors of the character*/
  voiceActors: Maybe<Staff[]>;
  /** The voice actors of the character with role date*/
  voiceActorRoles: Maybe<StaffRoleType[]>;
  /** The order the media should be displayed from the users favourites*/
  favouriteOrder: Maybe<number>;
}

/** Anime or Manga */
export interface Media {
  /** The id of the media*/
  id: number;
  /** The mal id of the media*/
  idMal: Maybe<number>;
  /** The official titles of the media in various languages*/
  title: Maybe<MediaTitle>;
  /** The type of the media; anime or manga*/
  type: Maybe<MediaType>;
  /** The format the media was released in*/
  format: Maybe<MediaFormat>;
  /** The current releasing status of the media*/
  status: Maybe<MediaStatus>;
  /** A media's statistics*/
  stats?: MediaStats;
  /** Short description of the media's story and characters*/
  description: Maybe<string>;
  /** The first official release date of the media*/
  startDate: Maybe<FuzzyDate>;
  /** The last official release date of the media*/
  endDate: Maybe<FuzzyDate>;
  /** The season the media was initially released in*/
  season: Maybe<MediaSeason>;
  /** The season year the media was initially released in*/
  seasonYear: Maybe<number>;
  /** The year & season the media was initially released in*/
  seasonInt: Maybe<number>;
  /** The amount of episodes the anime has when complete*/
  episodes: Maybe<number>;
  /** The general length of each anime episode in minutes*/
  duration: Maybe<number>;
  /** The amount of chapters the manga has when complete*/
  chapters: Maybe<number>;
  /** The amount of volumes the manga has when complete*/
  volumes: Maybe<number>;
  /** Where the media was created. (ISO 3166-1 alpha-2)*/
  countryOfOrigin: Maybe<string>;
  /** If the media is officially licensed or a self-published doujin release*/
  isLicensed: Maybe<boolean>;
  /** Source type the media was adapted from.*/
  source: Maybe<MediaSource>;
  /** Official Twitter hashtags for the media*/
  hashtag: Maybe<string>;
  /** Media trailer or advertisement*/
  trailer: Maybe<MediaTrailer>;
  /** When the media's data was last updated*/
  updatedAt: Maybe<number>;
  /** The cover images of the media*/
  coverImage: Maybe<MediaCoverImage>;
  /** The banner image of the media*/
  bannerImage: Maybe<string>;
  /** The genres of the media*/
  genres: Maybe<string[]>;
  /** Alternative titles of the media*/
  synonyms: Maybe<string[]>;
  /** A weighted average score of all the user's scores of the media*/
  averageScore: Maybe<number>;
  /** Mean score of all the user's scores of the media*/
  meanScore: Maybe<number>;
  /** The number of users with the media on their list*/
  popularity: Maybe<number>;
  /** Locked media may not be added to lists our favorited. This may be due to the entry pending for deletion or other reasons.*/
  isLocked: Maybe<boolean>;
  /** The amount of related activity in the past hour*/
  trending: Maybe<number>;
  /** The amount of user's who have favourited the media*/
  favourites: Maybe<number>;
  /** List of tags that describes elements and themes of the media*/
  tags: Maybe<MediaTag[]>;
  /** Other media in the same or connecting franchise*/
  relations: Maybe<MediaConnection>;
  /** The characters in the media*/
  characters: Maybe<CharacterConnection>;
  /** The staff who produced the media*/
  staff: Maybe<StaffConnection>;
  /** The companies who produced the media*/
  studios: Maybe<StudioConnection>;
  /** If the media is marked as favourite by the current authenticated user*/
  isFavourite: boolean;
  /** If the media is blocked from being added to favourites*/
  isFavouriteBlocked: boolean;
  /** If the media is intended only for 18+ adult audiences*/
  isAdult: Maybe<boolean>;
  /** The media's next episode airing schedule*/
  nextAiringEpisode: Maybe<AiringSchedule>;
  /** The media's entire airing schedule*/
  airingSchedule: Maybe<AiringScheduleConnection>;
  /** The media's daily trend stats*/
  trends: Maybe<MediaTrendConnection>;
  /** External links to another site related to the media*/
  externalLinks: Maybe<MediaExternalLink[]>;
  /** Data and links to legal streaming episodes on external sites*/
  streamingEpisodes: Maybe<MediaStreamingEpisode[]>;
  /** The ranking of the media in a particular time span and format compared to other media*/
  rankings: Maybe<MediaRank[]>;
  /** The authenticated user's media list entry for the media*/
  mediaListEntry: Maybe<MediaList>;
  /** User recommendations for similar media*/
  recommendations: Maybe<RecommendationConnection>;
  /** The url for the media page on the AniList website*/
  siteUrl: Maybe<string>;
  /** If the media should have forum thread automatically created for it on airing episode release*/
  autoCreateForumThread: Maybe<boolean>;
  /** If the media is blocked from being recommended to/from*/
  isRecommendationBlocked: Maybe<boolean>;
  /** If the media is blocked from being reviewed*/
  isReviewBlocked: Maybe<boolean>;
  /** Notes for site moderators*/
  modNotes: Maybe<string>;

  translations?: Translation[];
}
// ...
export interface FuzzyDate {
  /** Numeric Year (2017)*/
  year: Maybe<number>;
  /** Numeric Month (3)*/
  month: Maybe<number>;
  /** Numeric Day (24)*/
  day: Maybe<number>;
}
// ...
export interface CharacterConnection {
  edges: Maybe<CharacterEdge[]>;
  nodes: Maybe<Character[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}

/** Character connection edge */
export interface CharacterEdge {
  node: Maybe<Character>;
  /** The id of the connection*/
  id: Maybe<number>;
  /** The characters role in the media*/
  role: Maybe<CharacterRole>;
  /** Media specific character name*/
  name: Maybe<string>;
  /** The voice actors of the character*/
  voiceActors: Maybe<Staff[]>;
  /** The voice actors of the character with role date*/
  voiceActorRoles: Maybe<StaffRoleType[]>;
  /** The media the character is in*/
  media: Maybe<Media[]>;
  /** The order the character should be displayed from the users favourites*/
  favouriteOrder: Maybe<number>;
}
// ...
export enum StaffLanguage {
  Japanese = "JAPANESE",
  English = "ENGLISH",
  Korean = "KOREAN",
  Italian = "ITALIAN",
  Spanish = "SPANISH",
  Portuguese = "PORTUGUESE",
  French = "FRENCH",
  German = "GERMAN",
  Hebrew = "HEBREW",
  Hungarian = "HUNGARIAN",
}
/** Staff sort enums */
export enum StaffSort {
  Id = "ID",
  Id_desc = "ID_DESC",
  Role = "ROLE",
  Role_desc = "ROLE_DESC",
  Language = "LANGUAGE",
  Language_desc = "LANGUAGE_DESC",
  Search_match = "SEARCH_MATCH",
  Favourites = "FAVOURITES",
  Favourites_desc = "FAVOURITES_DESC",
  Relevance = "RELEVANCE",
}
/** Voice actors or production staff */
export interface Staff {
  /** The id of the staff member*/
  id: number;
  /** The names of the staff member*/
  name: Maybe<StaffName>;
  /** @deprecated Replaced with languageV2The primary language the staff member dub's in*/
  language: Maybe<StaffLanguage>;
  /** The primary language of the staff member. Current values: Japanese, English, Korean, Italian, Spanish, Portuguese, French, German, Hebrew, Hungarian, Chinese, Arabic, Filipino, Catalan, Finnish, Turkish, Dutch, Swedish, Thai, Tagalog, Malaysian, Indonesian, Vietnamese, Nepali, Hindi, Urdu*/
  languageV2: Maybe<string>;
  /** The staff images*/
  image: Maybe<StaffImage>;
  /** A general description of the staff member*/
  description: Maybe<string>;
  /** The person's primary occupations*/
  primaryOccupations: Maybe<string[]>;
  /** The staff's gender. Usually Male, Female, or Non-binary but can be any string.*/
  gender: Maybe<string>;
  dateOfBirth: Maybe<FuzzyDate>;
  dateOfDeath: Maybe<FuzzyDate>;
  /** The person's age in years*/
  age: Maybe<number>;
  /** [startYear, endYear] (If the 2nd value is not present staff is still active)*/
  yearsActive: Maybe<number[]>;
  /** The persons birthplace or hometown*/
  homeTown: Maybe<string>;
  /** The persons blood type*/
  bloodType: Maybe<string>;
  /** If the staff member is marked as favourite by the currently authenticated user*/
  isFavourite: boolean;
  /** If the staff member is blocked from being added to favourites*/
  isFavouriteBlocked: boolean;
  /** The url for the staff page on the AniList website*/
  siteUrl: Maybe<string>;
  /** Media where the staff member has a production role*/
  staffMedia: Maybe<MediaConnection>;
  /** Characters voiced by the actor*/
  characters: Maybe<CharacterConnection>;
  /** Media the actor voiced characters in. (Same data as characters with media as node instead of characters)*/
  characterMedia: Maybe<MediaConnection>;
  /** @deprecated No data available*/
  updatedAt: Maybe<number>;
  /** Staff member that the submission is referencing*/
  staff: Maybe<Staff>;
  /** Status of the submission*/
  submissionStatus: Maybe<number>;
  /** Inner details of submission status*/
  submissionNotes: Maybe<string>;
  /** The amount of user's who have favourited the staff member*/
  favourites: Maybe<number>;
  /** Notes for site moderators*/
  modNotes: Maybe<string>;
}

/** The names of the staff member */
export interface StaffName {
  /** The person's given name*/
  first: Maybe<string>;
  /** The person's middle name*/
  middle: Maybe<string>;
  /** The person's surname*/
  last: Maybe<string>;
  /** The person's first and last name*/
  full: Maybe<string>;
  /** The person's full name in their native language*/
  native: Maybe<string>;
  /** Other names the staff member might be referred to as (pen names)*/
  alternative: Maybe<string[]>;
  /** The currently authenticated users preferred name language. Default romaji for non-authenticated*/
  userPreferred: Maybe<string>;
}

export interface StaffImage {
  /** The person's image of media at its largest size*/
  large: Maybe<string>;
  /** The person's image of media at medium size*/
  medium: Maybe<string>;
}
// ...
export interface StaffArgs {
  /** Filter by the staff id*/
  id?: number;
  /** Filter by staff by if its their birthday today*/
  isBirthday?: boolean;
  /** Filter by search query*/
  search?: string;
  /** Filter by the staff id*/
  id_not?: number;
  /** Filter by the staff id*/
  id_in?: number[];
  /** Filter by the staff id*/
  id_not_in?: number[];
  /** The order the results will be returned in*/
  sort?: StaffSort[];
}